# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kürsü API",
    "description": "Anonim ders ve akademisyen değerlendirme platformunun API'si. Zarflar, akışlar, maskeleme ve hata gövdeleri gibi bu şemadan okunamayan kurallar için `docs/api-contract.md`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Create Review",
        "operationId": "create_review_reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_reviews_get",
        "parameters": [
          {
            "name": "course_professor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Professor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReviewResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/me": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List My Reviews",
        "operationId": "list_my_reviews_reviews_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReviewFullResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{review_id}": {
      "patch": {
        "tags": [
          "reviews"
        ],
        "summary": "Update My Review",
        "operationId": "update_my_review_reviews__review_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reviews"
        ],
        "summary": "Delete Review",
        "operationId": "delete_review_reviews__review_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/pending": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Pending Reviews",
        "operationId": "list_pending_reviews_reviews_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_professor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Professor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReviewFullResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{review_id}/status": {
      "patch": {
        "tags": [
          "reviews"
        ],
        "summary": "Update Review Status",
        "description": "Review'un KENDİSİNİN onayı/reddi. Bekleyen edit'in kararı ayrı uçta (/edit-status);\nburada reject, edit beklese bile review'u komple yayından kaldırır.",
        "operationId": "update_review_status_reviews__review_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{review_id}/edit-status": {
      "patch": {
        "tags": [
          "reviews"
        ],
        "summary": "Update Review Edit Status",
        "description": "Bekleyen edit'in onayı/reddi. approve gölge alanları asıl alanlara kopyalar,\nreject sadece temizler; iki durumda da review.status'a dokunulmaz (approved kalır).",
        "operationId": "update_review_edit_status_reviews__review_id__edit_status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ReviewFullResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewFullResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReviewFullResponse]"
      },
      "Page_ReviewResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReviewResponse]"
      },
...
      "ReviewCreate": {
        "properties": {
          "course_professor_id": {
            "type": "integer",
            "title": "Course Professor Id"
          },
          "teaching_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "course_professor_id",
          "teaching_score",
          "difficulty_score",
          "fairness_score"
        ],
        "title": "ReviewCreate"
      },
      "ReviewFullResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_professor_id": {
            "type": "integer",
            "title": "Course Professor Id"
          },
          "teaching_score": {
            "type": "integer",
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "has_pending_edit": {
            "type": "boolean",
            "title": "Has Pending Edit"
          },
          "pending_teaching_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Teaching Score"
          },
          "pending_difficulty_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Difficulty Score"
          },
          "pending_fairness_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Fairness Score"
          },
          "pending_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_professor_id",
          "teaching_score",
          "difficulty_score",
          "fairness_score",
          "comment",
          "status",
          "created_at",
          "has_pending_edit",
          "pending_teaching_score",
          "pending_difficulty_score",
          "pending_fairness_score",
          "pending_comment"
        ],
        "title": "ReviewFullResponse"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_professor_id": {
            "type": "integer",
            "title": "Course Professor Id"
          },
          "teaching_score": {
            "type": "integer",
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_professor_id",
          "teaching_score",
          "difficulty_score",
          "fairness_score",
          "comment",
          "status",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
      "ReviewStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ReviewStatusUpdate"
      },
      "ReviewUpdate": {
        "properties": {
          "teaching_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "teaching_score",
          "difficulty_score",
          "fairness_score"
        ],
        "title": "ReviewUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}